Change the title and update the difference
Add comments on clarify the difference among pointer constant, constant pointer, and constant pointer points to constant.
